+++ packages/logger/src/formatters/logger-formatter-file.ts
@@ -63,7 +63,8 @@ export function formatLoggerFileEntry(entry: LoggerEntry): string {
   const message = `[${date}] [${level}] [${entry.package}] [${entry.appName}]${statusCode} - ${entry.message}`;
 
   const context = formatLoggerContext(
-    entry.context
+    entry.context,
+    entry.verboseError
   );
 
   if (!context) {